+++ src/web/assets/cp/src/plugin-settings/js/formie-plugin-settings.js
@@ -8,5 +8,7 @@
 // CSS needs to be imported here as it's treated as a module
 import '../scss/formie-plugin-settings.scss';
 import { initDurationHints } from './duration-hint';
+import { initCaptchaIntegrationStatus } from './captcha-integration-status';
 
 initDurationHints();
+initCaptchaIntegrationStatus();